+++ code/src/components/experiments.js
@@ -1,17 +1,17 @@
 import React from 'react';
 
 import Experiment from './experiment';
-import { projects } from './projects';
+import { projects } from '../utils/projects';
+import { shuffle } from '../utils/generic';
 
 const Experiments = () => {
-    // const shuffle = (array) => array.sort(() => Math.random() - 0.5);
-    // const shuffled = shuffle(projects);
+    const shuffled = shuffle(projects);
     return (
         <section className="section-projects spacer-in-top-xxl">
             <h2 className="section-title section-title--underlined">Experiments</h2>
             <div className="row list-project">
                 {
-                    projects.map((project) => 
+                    shuffled.map((project) => 
                     <Experiment 
                             key={ project.id }
                             cover = { project.cover }
